+++ NEWS.md
@@ -1,5 +1,8 @@
 # pkgdown 1.1.0.9000
 
+* `build_reference_index()`: Selectors that do not match topics now generate a warning.
+  If none of the specified selectors have a match, no topics are selected (#728).
+  
 * Support of multiple arguments in `\Sexpr{}` was fixed, eliminating `x must be a
   string or a R connection` errors when using `\doi` Rd tags (#738).